Fix Curve and Balancer quoters
- Add tests

+++ src/portals/quoter/interface/ICurvePool.sol
@@ -12,5 +12,8 @@ interface ICurvePool {
         view
         returns (uint256 out);
 
-    function coins(int128 arg0) external view returns (address out);
+    function coins(uint256 arg0)
+        external
+        view
+        returns (address out);
 }
